Refactored, no warnings for long to int

+++ Garden_messaingaround/Garden_messaingaround/ViewControllerEditAlarm.h
@@ -19,7 +19,7 @@
 @property (strong, nonatomic) PlantObject * plant ;
 @property ViewControllerEditAlarm * reminders;
 @property NSString * selectedPlant;
-@property int plantid;
+@property NSInteger plantid;
 @property BOOL repeat;
 
 @end
